Move Dao files to a Dao folder.
This change was made to make the project structure cleaner. Additionally, if more Daos were to be implemented in the future. They would be in this folder.
