+++ GoMeet/src/main/java/com/google/sps/servlets/LocationServlet.java
@@ -1,9 +1,5 @@
 package main.java.com.google.sps.servlets;
 
-import main.java.com.google.sps.data.Dao;
-import main.java.com.google.sps.data.Location;
-import main.java.com.google.sps.data.LocationDao;
-
 import com.google.gson.Gson;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -13,6 +9,9 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import main.java.com.google.sps.dao.Dao;
+import main.java.com.google.sps.dao.LocationDao;
+import main.java.com.google.sps.data.Location;
 import org.jsoup.Jsoup;
 import org.jsoup.safety.Whitelist;
