+++ GoMeet/src/test/java/com/google/sps/LocationServletTest.java
@@ -17,9 +17,8 @@ import com.google.appengine.api.datastore.Key;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 import main.java.com.google.sps.servlets.LocationServlet;
+import main.java.com.google.sps.dao.LocationDao;
 import main.java.com.google.sps.data.Location;
-import main.java.com.google.sps.data.LocationDao;
-import java.beans.Transient;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
